refactor(storage-serializer): reworked the serializers; corrected previous issue with unserializing php
Refs #34
