+++ src/Storage/Serializer/Json.php
@@ -14,14 +14,11 @@ declare(strict_types=1);
 namespace Phalcon\Storage\Serializer;
 
 use InvalidArgumentException;
-use JsonException;
 use JsonSerializable;
 
 use function is_object;
 use function json_decode;
 use function json_encode;
-
-use const JSON_THROW_ON_ERROR;
 
 /**
  * Class Json
@@ -33,19 +30,22 @@ class Json extends AbstractSerializer
     /**
      * Serializes data
      *
-     * @return false|JsonSerializable|mixed|string|null
-     * @throws JsonException
+     * @return JsonSerializable|mixed|string
      */
     public function serialize()
     {
-        if (is_object($this->data) && !($this->data instanceof JsonSerializable)) {
+        if (true === is_object($this->data) && !($this->data instanceof JsonSerializable)) {
             throw new InvalidArgumentException(
                 'Data for the JSON serializer cannot be of type "object" ' .
                 'without implementing "JsonSerializable"'
             );
         }
 
-        return parent::serialize();
+        if (true !== $this->isSerializable($this->data)) {
+            return $this->data;
+        }
+
+        return json_encode($this->data);
     }
 
     /**
@@ -53,37 +53,10 @@ class Json extends AbstractSerializer
      *
      * @param string $data
      *
-     * @throws JsonException
+     * @return void
      */
-    public function unserialize($data): void
+    public function unserialize($data)
     {
-        $this->data = $this->internalUnserlialize($data);
-    }
-
-    /**
-     * @param mixed $data
-     *
-     * @return false|string
-     * @throws JsonException
-     */
-    protected function internalSerialize($data)
-    {
-        return json_encode($this->data, 79 + JSON_THROW_ON_ERROR);
-    }
-
-    /**
-     * @param mixed $data
-     *
-     * @return mixed
-     * @throws JsonException
-     */
-    protected function internalUnserlialize($data)
-    {
-        return json_decode(
-            $data,
-            false,
-            512,
-            JSON_THROW_ON_ERROR
-        );
+        $this->data = json_decode($data);
     }
 }
