+++ src/Storage/Serializer/Msgpack.php
@@ -13,6 +13,13 @@ declare(strict_types=1);
 
 namespace Phalcon\Storage\Serializer;
 
+use function msgpack_pack;
+use function msgpack_unpack;
+use function restore_error_handler;
+use function set_error_handler;
+
+use const E_WARNING;
+
 /**
  * Class Msgpack
  *
@@ -21,22 +28,44 @@ namespace Phalcon\Storage\Serializer;
 class Msgpack extends AbstractSerializer
 {
     /**
-     * @param mixed $data
+     * Serializes data
      *
-     * @return mixed
+     * @return string|null
      */
-    protected function internalSerialize($data)
+    public function serialize()
     {
-        return msgpack_pack($data);
+        if (true !== $this->isSerializable($this->data)) {
+            return $this->data;
+        }
+
+        return msgpack_pack($this->data);
     }
 
     /**
-     * @param mixed $data
+     * Unserializes data
      *
-     * @return mixed
+     * @param string $data
+     *
+     * @return void
      */
-    protected function internalUnserlialize($data)
+    public function unserialize($data)
     {
-        return msgpack_unpack($data);
+        $warning = false;
+        set_error_handler(
+            function () use (&$warning) {
+                $warning = true;
+            },
+            E_WARNING
+        );
+
+        $data = msgpack_unpack($data);
+
+        restore_error_handler();
+
+        if (true === $warning) {
+            $data = null;
+        }
+
+        $this->data = $data;
     }
 }
